Fix GT tool material being ignored
